+++ Lib/fontbakery/profiles/universal.py
@@ -21,6 +21,7 @@ PROFILE = {
             "ufo_unnecessary_fields",
         ],
         "Universal Profile Checks": [
+            "allah_ligature",
             "alt_caron",
             "arabic_high_hamza",
             "arabic_spacing_symbols",